hashed passwords in db + send token after login

+++ customnews/src/main/java/com/frontpagenews/controllers/ArticleController.java
@@ -1,8 +1,12 @@
 package com.frontpagenews.controllers;
 
+import com.frontpagenews.models.AdminModel;
 import com.frontpagenews.models.ArticleModel;
+import com.frontpagenews.services.AdminService;
 import com.frontpagenews.services.ArticleService;
 
+import com.mongodb.util.JSON;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,14 +24,24 @@ public class ArticleController {
     @Autowired
     ArticleService articleService;
 
+    @Autowired
+    AdminService adminService;
+
     @RequestMapping(method= RequestMethod.GET)
     public List<ArticleModel> getAllArticles() {
         return articleService.getAll();
     }
 
     @RequestMapping(method=RequestMethod.POST)
-    public ArticleModel createArticle(@Valid @RequestBody ArticleModel admin) {
-        return articleService.save(admin);
+    public ResponseEntity<String> createArticle(@Valid @RequestBody ArticleModel article, @RequestHeader("Token") String token) {
+        AdminModel admin = adminService.getByToken(token);
+        if (admin != null) {
+            articleService.save(article);
+            return new ResponseEntity<String>("{\"OK\": 1}", HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<String>("{\"OK\": 1, \"error\": \"Permission denied\"}", HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping(value="/page/{page}/{language}", method=RequestMethod.GET)
@@ -86,10 +100,10 @@ public class ArticleController {
     }
 
     @RequestMapping(value="{id}", method=RequestMethod.PUT)
-    public ResponseEntity<ArticleModel> updateArticle(@Valid @RequestBody ArticleModel article, @PathVariable("id") String id) {
+    public ResponseEntity<String> updateArticle(@Valid @RequestBody ArticleModel article, @PathVariable("id") String id, @RequestHeader("Token") String token) {
         ArticleModel articleData = articleService.getById(id);
         if(articleData == null) {
-            return new ResponseEntity<ArticleModel>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<String>("{\"OK\": 1, \"error\": \"Article doesn't exist\"}", HttpStatus.NOT_FOUND);
         }
         articleData.setTitle(article.getTitle());
         articleData.setContent(article.getContent());
@@ -103,13 +117,30 @@ public class ArticleController {
         articleData.setImageWidth(article.getImageWidth());
         articleData.setSummary(article.getSummary());
         articleData.setLanguage(article.getLanguage());
-        ArticleModel updatedArticle = articleService.save(articleData);
-        return new ResponseEntity<ArticleModel>(updatedArticle, HttpStatus.OK);
+        AdminModel admin = adminService.getByToken(token);
+        if (admin != null) {
+            articleService.save(articleData);
+            return new ResponseEntity<String>("{\"OK\": 1}", HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<String>("{\"OK\": 1, \"error\": \"Permission denied\"}", HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping(value="{id}", method=RequestMethod.DELETE)
-    public void deleteArticle(@PathVariable("id") String id) {
-        articleService.delete(id);
+    public ResponseEntity<String> deleteArticle(@PathVariable("id") String id, @RequestHeader("Token") String token) {
+        ArticleModel article = articleService.getById(id);
+        if(article == null) {
+            return new ResponseEntity<String>("{\"OK\": 1, \"error\": \"Article doesn't exist\"}", HttpStatus.NOT_FOUND);
+        }
+        AdminModel admin = adminService.getByToken(token);
+        if (admin != null) {
+            articleService.delete(id);
+            return new ResponseEntity<String>("{\"OK\": 1}", HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<String>("{\"OK\": 1, \"error\": \"Permission denied\"}", HttpStatus.UNAUTHORIZED);
+        }
     }
     
 }
